refactor(template): factor `Checksum` out into function
